feat: Read Kinvey keys from package.json
The new version of kinvey-nativescript-sdk allows calling `Kinvey.init()` without arguments. The plugin reads the data from app/package.json internally and uses the values for appSecret and appKey.
So remove the config.ts file and add instructions how to set the data in package.json. Predefine empty values for the keys, so when the template is installed in Sidekick, we will be able to determine that the plugin configuration is not correct and prompt the user to create Kinvey app.